+++ src/main/java/org/springframework/samples/petclinic/sector/SectorController.java
@@ -85,8 +85,11 @@ public class SectorController {
 
     @DeleteMapping()
     public ResponseEntity<Void> deleteSectorsByGameId(
-            @ParameterObject @RequestParam(value = "gameid", required = false) Integer gameid) {
-        scs.deleteByGameId(gameid);
+            @ParameterObject @RequestParam(value = "gameid", required = true) Integer gameid) {
+        if (getSectorById(gameid) == null){
+            return ResponseEntity.notFound().build();
+        }
+            scs.deleteByGameId(gameid);
         return ResponseEntity.noContent().build();
     }
